requirements.txt
faiss-cpu
sentence-transformers
transformers
fastapi
uvicorn[standard]